ShibIntIdP3: remove IdPv2 API dependencies
Remove IdPv2 API calls no longer available in IdPv3.

Get relyingparty from request attributes instead.

NOTE: this is not available in a filter and this will always return a
null value.  (Even though the VHR happily ignores the null value).

This gets the code to compile with IdPv3.

NOTE: this will not work yet, as the query string is lost during the
direct to VHR and back and the webflow looses its state.