ShibIntIdP3: add VhrRemoteUserAuthServlet
Add VhrRemoteUserAuthServlet as the proper solution for integrating
IdPv3 with VHR.

The profile context references needed for proper interaction with IdPv3
are only available to a Servlet, not to a Filter.

Abandon VhrFilter and instead start developing a Servlet with the same
functionality.

Start with an identical copy of VhrRemoteUserAuthServlet, with only
minimal initial bootstrapping changes:
* class and package renamed (incl. internal references to class itself)
* serialVersionUID updated (to what future state with attributes to be
* imported from VhrFilter will be)
* ... and trailing whitespace removed