ShibIntIdP3: VhrRUAS: reorganize code
Reorganize method code to explicitly support two flows: initial login
and redirect from VHR.

Add a request parameter to be used in the return URL to indicate this is
a redirect back from the VHR.  And drop the current query string
(conversation key) from the return URL - we will instead rely on
ExternalAuthentication.finishExternalAuthentication to pick up where the
flow finished.

In the reorganized code, we need to cover these use cases:

Regular (Success) use cases:
* initial login; no cookie => redirect to VHR
* returned from VHR with a cookie => validate and complete authentication

Error use cases:
* returned from VHR but no cookie => error
* returned from VHR; cookie found but failed to validate => error

* Initial flow: cookie found but failed to validate.  This might be the
  edge-case of just expiring auth session at VHR (browser still included
  the cookie, but VHR failed to validate).  Could be legit flow (with
  some time fuzz) => redirected to VHR

This simplifies the code: avoids duplication of the redirect code.

If the user has returned from VHR, we do not redirect there again
(and thus we do not need relyingParty which we would not see in the
request attributes)