ShibIntIdP3: support forceAuthn in authn request
SAML 2.0 allows an SP to request forced new authentication (interaction
with Subject) by including forceAuthn="true" in the SSO request.

IdPv3 passes this to the ExternalAuthentication servlet in the (Boolean)
forceAuthn attribute on the initial HTTP request.

This had been unsupported by the VHR integration module.

With IdP 2.x, the forceAuthn SSO request parameter was plain ignored.

With IdP 3.x, the default configuration does not mark RemoteUser
authentication as supporting forcedAuthentication, and that makes the
IdP fail the SSO request.

To declare support for forceAuthn, we need to first properly implement
support.

This implementation includes:
* Extracting forceAuthn from the request attribute.
* Extracting the instant the authentication started at the IdP from the request attribute.
* Modifying VhrSessionValidator to also extract the authnInstant out of the VHR response.
  * This relies on a corresponding change to VHR itself to return the authnInstant in the API response.
* Modifying VhrSessionValidator to compare the authnInstant at the VHR to a "notBefore" cut-off threshold.
* In case of forceAuthn, pass the authnStart instant at the IdP to VhrSessionValidator as the notBefore cut-off threshold.
* Preserving the isForceAuthn and authnStart values across redirects to the VHR (in httpSession attributes)

The above rules guarantee that with forceAuthn, VhrSessionValidator will
only accept authentication from the VHR with an authnInstant *after* the
authentication started at the IdP - so fresh authentication.

To declare support for forceAuthn, make the corresponding change to
conf/authn/general-authn.xml as shown in
configuration/idpv3/change-in-authn-general-authn.xml

When running with a version of VHR not returning authnInstant,
forceAuthn SSO requests will fail.

This code adds dependency on joda-time (version 2.9, same as packaged with IdP 3.2.1).

In JSON, authnInstant is passed as string, formatted as ISO8601 with
seconds included down to three decimal places (milliseconds).