ShibIntIdP3: pass authnInstant back to IdP
IdPv3 allows an ExternalAuthentication servlet to pass back the
authentication instant in the (org.joda.time.DateTime) authnInstant
request attribute.

As this value is already received back from the VHR (to support
forceAuthn), return it back to the IdP.

* Modify VhrSessionValidator to pass the authnInstant back as an additional output parameter.
  * Pass a reference to an DateTime[] array of size 1 to hold the value.
  * Populate this vlaue if returned from VHR.
* Set this value as the authnInstant attribute on the request object if returned by VhrSessionValidator.

Note that this is actually not working with IdP 3.2.1, as the value is
not passed from the ExternalAuthenticationContext to the
AuthenticationContext.AuthenticationResult object.

This will be fixed in IdP 3.3.0 once released.

Please see https://issues.shibboleth.net/jira/browse/IDP-1031 for
further information.