Upgrade to AAF patched groovy 2.0.8 release
We've manually patched the Groovy version for Groovy 7664, this commit
brings it into the war for release.

Note fallback to hard delete of existing 2.0.8 binary, for whatever
reason the dependency resolution failed to do this on its own thus
I had no option but the sledgehammer approach to ensure only the patched
classes are in use.