Update to recaptcha v2
As recaptcha v1 has been deprecated requests which required this
functionality failed completely.

Our old version of Grails is still very problematic here but I was able
to get the recaptcha plugin updated and working for the v2 API, albeit
not the latest version of the plugin which is Grails 2.3+ only.

I've tested that this functions correctly on both lost username and
multiple failed logins.

Bandaids and sticky tape to the rescue!.